test: centralize agent_bin() in test-support, add llvm-cov-target lookup
Eight loopback tests had a copy-pasted agent_bin() helper that walked
target/{debug,release} for reef-agent. The cargo-llvm-cov coverage CI
job actually puts the binary in target/llvm-cov-target/debug/, and
contrary to the previous fix attempt's assumption, cargo-llvm-cov sets
that via --target-dir flag rather than the CARGO_TARGET_DIR env var,
so the runtime env-var probe never matched.

Move the helper to test-support and have it look at all three layouts
(CARGO_TARGET_DIR, target/llvm-cov-target, target). Also revert the
ci.yml --all-targets workaround since reef-agent was always being
built; the failure was purely the lookup, not the build.
